+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-Release *v2.0.dev5* - ``2019-10-07``
+Release *v2.0.dev6* - ``2019-10-07``
 ------------------------------------
 * Support for redis-py >= 3 only
 * Support for redis-server >= 3 only
@@ -15,6 +15,7 @@ Release *v2.0.dev5* - ``2019-10-07``
 * Add `setex` and `psetex` to `StringField`. Can only be called on non-indexable fields.
 * Deny `ex` and `px` flag to `set` if field is indexable
 * Add support for `count` argument to `spop` (only for redis-server >= 3.2)
+* Add new commands to `ListField`: `lcontains`, `lrank` and `lcount`, to know if a value is in the list, where, and how many times. This is done on the redis server side via lua scripting.
 
 Release *v1.3.1* - ``2019-10-11``
 ---------------------------------